night theme working, may break at midnight
