+++ plugin-interface/src/components/TimeTheme/TimeTheme.js
@@ -5,26 +5,17 @@ const axios = require('axios');
 
 
 const TimeTheme = ({ key, manager }) => {
-    const [time, setTime] = useState(new Date());
-    const [location, setLocation] = useState(null);
-    const [sunset, setSunset] = useState(null);
-    const [sunsetDate, setSunsetDate] = useState(new Date());
-    const [sunsetHour, setSunsetHour] = useState(null);
-    const [sunsetMinute, setSunsetMinute] = useState(null);
-    const [sunrise, setSunrise] = useState(null);
-    const [sunriseDate, setSunriseDate] = useState(new Date());
-    const [isItSunset, setIsItSunset] = useState(false);
-    const [deBug, setDeBug] = useState([]);
-    const [deBugRise, setDeBugRise] = useState([]);
-    const [isItSunrise, setIsItSunrise] = useState(false);
-    const [isItDay, setIsItDay] = useState(false);
 
-    const hours = time.getHours();
-    const minutes = time.getMinutes();
-    const seconds = time.getSeconds();
-    const offset = time.getTimezoneOffset() / 60;
-    const adjust = time.getTime() + (8 * 60 * 60 * 1000);
     //const adjust = 0;
+    const [location, setLocation] = useState(null);
+    const [currentTime, setCurrentTime] = useState(new Date());
+    const [debug, setDebug] = useState([]);
+    const [apiData, setApiData] = useState({});
+    const [sunset, setSunset] = useState(null);
+    const [sunrise, setSunrise] = useState(null);
+    const [dayLength, setDayLength] = useState(null);
+    const [sunSetTime, setSunSetTime] = useState(new Date());
+    const [isDayTime, setIsDayTime] = useState(false);
 
 
     //get lattitude and longitude
@@ -34,159 +25,106 @@ const TimeTheme = ({ key, manager }) => {
         setLocation(response.data);
         return response.data;
     }
-    const findSunset = (sunny) => {
-        const sunsetTime = sunny.results.sunset.split(':');
-        const sunsetHour = parseInt(sunsetTime[0]);
-        const sunsetMinute = parseInt(sunsetTime[1]);
-        const sunsetOffset = sunsetHour - offset;
-        const sunsetDate = new Date(time.getFullYear(), time.getMonth(), time.getDate(), sunsetOffset, sunsetMinute);
-        if (sunsetDate.getTime() >= time.getTime()) {
 
-            setIsItSunset(true);
-        }
-        else {
-            setIsItSunset(false);
-
-        }
-        setDeBug([sunsetDate.getTime(), time.getTime(), time.getTime() - sunsetDate.getTime()]);
-        return sunsetDate;
-    }
-    /*
-    START NEW CODE
-    */
-    const nDate = (time) => {
-        let sunriseTime = time.split(':');
-        let sunriseHour = parseInt(sunriseTime[0]);
-        console.log("Hour: " + sunriseHour);
-        let sunriseMinute = parseInt(sunriseTime[1]);
-        const meridian = sunriseTime[2].split(' ');
-        const fakeDate = new Date();
-        if (meridian[1] === 'PM') {
-            sunriseHour += 12;
-        }
-        const retDate = new Date(Date.UTC(fakeDate.getFullYear(), fakeDate.getMonth(), fakeDate.getUTCDate(), sunriseHour, sunriseMinute, 0, 0));
-
-        return retDate;
-    }
-    const ssDate = (time) => {
-        //sunrise time
-        let sunriseTime = time.split(':');
-        let sunriseHour = parseInt(sunriseTime[0]);
-        console.log("Sunrise Hour: " + sunriseHour);
-        let sunriseMinute = parseInt(sunriseTime[1]);
-        const meridian = sunriseTime[2].split(' ');
-        const fakeDate = new Date();
-        if (meridian[1] === 'PM') {
-            sunriseHour += 12;
-        }
-        const retDate = new Date(Date.UTC(fakeDate.getFullYear(), fakeDate.getMonth(), fakeDate.getUTCDate() + 1, sunriseHour, sunriseMinute, 0, 0));
-
-        return retDate;
-    }
-    const qDate = () => {
-        const nowDate = new Date();
-        const nowDateUTC = new Date(Date.UTC(nowDate.getFullYear(), nowDate.getMonth(), nowDate.getDate(), nowDate.getHours(), nowDate.getMinutes(), nowDate.getSeconds(), nowDate.getMilliseconds()));
-        return nowDate;
-    }
-    // const sunsetXDate = ssDate(results.sunset);
-    // const sunriseXDate = nDate(results.sunrise);
-    // const nowxDate = qDate();
-
-
-
-    /*
-    END NEW CODE
-
-    */
-    const findSunrise = (sunny) => {
-        const sunriseTime = sunny.results.sunrise.split(':');
-        const sunriseHour = parseInt(sunriseTime[0] - offset);
-        if (sunny.results.sunrise.indexOf("PM") > -1) {
-            const sunriseHour = parseInt(sunriseTime[0]);
-        }
-
-        const sunriseMinute = parseInt(sunriseTime[1]);
-        const sunriseOffset = sunriseHour - offset;
-        const sunriseDate = new Date(time.getFullYear(), time.getMonth(), time.getDate(), sunriseOffset, sunriseMinute);
-        // if (sunriseDate.getTime() <= time.getTime()) {
-        //     setIsItSunset(false);
-        // }
-        // else {
-        //     setIsItSunset(true);
-        // }
-        setDeBugRise([`Sunrise ${sunriseDate.getHours()} : ${sunriseDate.getMinutes()}`, `Now ${time.getHours()} : ${time.getMinutes()}`, `Sunset: ${sunsetDate.getHours()}:${sunsetDate.getMinutes()}`]);
-        return sunriseDate;
-    }
-    //rename this function
-    const getSunset = async (local) => {
+    const sunSetSunRiseApi = async (local) => {
         const response = await axios.get(`https://api.sunrise-sunset.org/json?lat=${local.latitude}&lng=${local.longitude}&date=today`);
         //  onlt really need *date
+        console.log("DEBUGHIT", response.data);
+        setApiData(response.data);
+        const processedSunRise = processSunRise(response.data.results.sunrise)
+        setDayLength(response.data.results.day_length);
         //https://api.sunrise-sunset.org/json?lat=34.1624&lng=-118.1275&date=today
-        setSunsetDate(ssDate((response.data.results.sunset)));
-        setSunriseDate(nDate(response.data.results.sunrise));
-        // setSunsetDate(ssDate(response.results.sunset));
-        // setSunriseDate(nDate(response.results.sunrise));
-        if (qDate().getTime() < sunsetDate.getTime()) {
-            setIsItDay(true);
+        const dayLength = processDayLength(response.data.results.day_length);
+        const sunset = processedSunRise + dayLength;
+        let sunSetTime = new Date();
+        sunSetTime.setTime(sunset);
+        setSunSetTime(sunSetTime);
+        if (currentTime.getTime() < sunSetTime) {
+            setIsDayTime(true);
             manager.updateConfig({ colorTheme: FeatherTheme });
-        }
-        else {
-            setIsItDay(false);
+        } else {
+            setIsDayTime(false);
         }
 
-        //34.1624 -118.1275
-        //https://api.sunrise-sunset.org/json?lat=34.1624&lng=-118.1275&date=today
+
         return response.data;
     }
+    function processSunRise(timeSunRise) {
+        let time = new Date();
+        const sunriseTime = timeSunRise.split(':');
+        let sunriseHour = parseInt(sunriseTime[0]);
+        const sunriseMinute = parseInt(sunriseTime[1]);
+        const meridian = timeSunRise.split(' ')[1];
+        if (meridian === 'PM') {
+            sunriseHour = sunriseHour + 12;
+        }
+        sunriseHour = sunriseHour - time.getTimezoneOffset() / 60;
+        //ADD TESTING HERE TO SEE IF IT IS DAY OR NIGHT
+        let sunDate = new Date(time.getFullYear(), time.getMonth(), time.getDate(), sunriseHour, sunriseMinute);
+        time.setTime(sunDate.getTime());
 
+        setSunrise(sunDate);
+        return sunDate.getTime();
+
+    }
+    function processDayLength(timeLength) {
+        const timeLengthArray = timeLength.split(':');
+        const hour = parseInt(timeLengthArray[0]);
+        const minute = parseInt(timeLengthArray[1]);
+        //combine hours and minutes to miliseconds
+        const dayLength = (hour * 60 * 60 * 1000) + (minute * 60 * 1000);
+
+        setDebug([dayLength, hour, minute]);
+
+        setDayLength(dayLength);
+        return dayLength;
+    }
+
+
+
+
+    //moment js
 
     useEffect(() => {
         getLocation()
             .then(local => {
-                getSunset(local);
+                setApiData(sunSetSunRiseApi(local));
+                console.log(apiData);
             })
-
     }, [])
-    useEffect(() => {
-    }, [sunset])
 
 
-    //get sunrise from axios call
-
-    useEffect(() => {
-
-        setInterval(() => {
-            setTime(new Date());
-            if (isItDay === true) {
-                // manager.updateConfig({ colorTheme: FeatherTheme });
-                // setDeBug([isItDay])
-            }
-        }, 1000);
-    }, []);
 
 
 
     const myName = "Vincent"
     return (
         <div>
-            Now Time: what the hell <strong>{`${hours}: ${minutes}:${seconds} -- ${offset}`}</strong>?
             <p>
                 {
-                    location ? `Lattitude: ${location.latitude} Longitude: ${location.longitude}` : null
+                    location ? `Lattitude: ${location.latitude} || Longitude: ${location.longitude}` : null
                 }
             </p>
             <p>
-
-
-            </p>
-            <p>
-
-            </p>
-            <p>
+                //map through apiData
+                {
+                    apiData.results ? `Sunrise: ${sunrise}` : null
+                }
 
             </p>
             <p>
-
+                {
+                    apiData.results ? `Sunset: ${sunSetTime}` : null
+                }
+            </p>
+            <p>
+                dayLength: {dayLength}
+            </p>
+            <p>
+                DEBUG {debug[0]} and {debug[1]}
+            </p>
+            <p>
+                Is Daylight {isDayTime}
             </p>
         </div>
     )
